+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+* ThreeDSecure
+  * Bump Cardinal version to `2.2.5-4`
 * SamsungPay
   * Add `BuildCustomSheetPaymentInfoCallback`
   * Add `SamsungPayClient`