+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * Make `Configuration#getSupportedCardTypes()` public
 * ThreeDSecure
   * Make `ThreeDSecureResult#getLookup()` public
+  * Bump Cardinal version to `2.2.5-4`
 * SamsungPay
   * Add `BuildCustomSheetPaymentInfoCallback`
   * Add `SamsungPayClient`